use a struct for totals
This refactor enables multiple modes of counting at a time. Also uses
the Display trait for formatting results.

TODO: When there are multiple counting modes, `wc` outputs numbers in
equally-padded columns. That functionality should be duplicated.

+++ src/bytes.rs
@@ -1,5 +1,5 @@
-pub fn count(text: &str) -> usize {
-    text.len()
+pub fn count(text: &str) -> u64 {
+    text.len() as u64
 }
 
 #[cfg(test)]
